requirements.txt
apache-beam==2.51.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cloudpickle==2.2.1
crcmod==1.7
dill==0.3.1.1
distlib==0.3.7
dnspython==2.4.2
docopt==0.6.2
fastavro==1.9.0
fasteners==0.19
filelock==3.13.0
google-api-core==2.12.0
google-apitools==0.5.31
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-aiplatform==1.36.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
google-cloud-bigtable           2.21.0
google-cloud-core               2.3.3
google-cloud-datastore          2.18.0
google-cloud-dlp                3.13.0
google-cloud-language           2.11.1
google-cloud-pubsub             2.18.4
google-cloud-pubsublite         1.8.3
google-cloud-recommendations-ai 0.10.5
google-cloud-resource-manager   1.10.4
google-cloud-spanner            3.40.1
google-cloud-storage            2.13.0
google-cloud-videointelligence  2.11.4
google-cloud-vision             3.4.5
google-crc32c                   1.5.0
google-resumable-media          2.6.0
googleapis-common-protos        1.61.0
grpc-google-iam-v1              0.12.6
grpcio                          1.59.2
grpcio-status                   1.59.2
hdfs                            2.7.3
httplib2                        0.22.0
idna                            3.4
Js2Py                           0.74
mysql                           0.0.3
mysql-connector                 2.2.9
mysql-connector-python          8.2.0
mysqlclient                     2.2.0
numpy                           1.24.4
oauth2client                    4.1.3
objsize                         0.6.1
orjson                          3.9.10
overrides                       6.5.0
packaging                       23.2
pandas                          2.1.2
pip                             23.2.1
pipenv                          2023.10.24
platformdirs                    3.11.0
proto-plus                      1.22.3
protobuf                        4.24.4
pyarrow                         11.0.0
pyasn1                          0.5.0
pyasn1-modules                  0.3.0
pydot                           1.4.2
pyjsparser                      2.7.1
pymongo                         4.6.0
pyparsing                       3.1.1
python-dateutil                 2.8.2
python-dotenv                   1.0.0
pytz                            2023.3.post1
regex                           2023.10.3
requests                        2.31.0
rsa                             4.9
schedule                        1.2.1
setuptools                      68.2.2
shapely                         2.0.2
six                             1.16.0
sqlparse                        0.4.4
typing_extensions               4.8.0
tzdata                          2023.3
tzlocal                         5.2
urllib3                         2.0.7
virtualenv                      20.24.6
zstandard                       0.22.0